change invalid period into valid comma
I have absolutely no idea what happened here. Libsass refused to import
this file, and I was outraged, because how dare sass claim to be too
good for tachyons-verbose! But then I went and looked at the file and it
is actually invalid css? This really seems like it was supposed to be a
comma and somehow became a period instead. Changing it makes sass happy,
too.